完成批量或单条删除课程信息功能
