+++ mooc-web-admin/src/main/resources/templates/course/course_list.html
@@ -128,7 +128,7 @@
                                 新增
                             </a>&nbsp;&nbsp;&nbsp;
                             <button type="button" class="btn btn-sm btn-default"
-                                    onclick="App.deleteMulti('/user/delete')"><i class="fa fa-trash-o"></i> 删除
+                                    onclick="App.deleteMulti('/course/delete')"><i class="fa fa-trash-o"></i> 删除
                             </button>&nbsp;&nbsp;&nbsp;
                             <button type="button" class="btn btn-sm btn-primary"
                                     onclick="$('.box-info-search').css('display') == 'none' ? $('.box-info-search').show('fast') : $('.box-info-search').hide('fast')">
@@ -196,7 +196,7 @@
             {
                 "data": function (row, type, val, meta) {
                     var detailUrl = "/course/detail?id=" + row.id;
-                    var deleteUrl = "/user/delete";
+                    var deleteUrl = "/course/delete";
                     return '<button type="button" class="btn btn-sm btn-default" onclick="App.showDetail(\'' + detailUrl + '\');"><i class="fa fa-search"></i> 查看</button>&nbsp;&nbsp;&nbsp;' +
                         '<a href="/course/form?id=' + row.id + '" type="button" class="btn btn-sm btn-primary"><i class="fa fa-edit"></i> 编辑</a>&nbsp;&nbsp;&nbsp;' +
                         '<button type="button" class="btn btn-sm btn-danger" onclick="App.deleteSingle(\'' + deleteUrl + '\', \'' + row.id + '\')"><i class="fa fa-trash-o"></i> 删除</button>';
